move credential creation from IcebergTableEntry up to IcebergTableInformation, use it in the 'iceberg_use_metadata_log' path, since we need to read the additional .json file, we need the right credentials
