+++ test/sql/local/catalog_test_config_setup/catalog_agnostic/delete/test_transactional_delete.test
@@ -135,7 +135,7 @@ select * from my_datalake.default.multi_transaction_insert order by all;
 statement error con1
 commit;
 ----
-<REGEX>:.*branch main has changed: expected id.*
+<REGEX>:.*Failed to commit.*
 
 # con1's DELETE was a no-op; con2's committed row remains.
 query II
