+++ test/sql/local/catalog_test_config_setup/catalog_agnostic/transactions/test_transaction_metadata_log_layout.test
@@ -58,7 +58,7 @@ a
 statement error con1
 commit
 ----
-<REGEX>:.*Cannot find source column for partition field.*
+<REGEX>:.*Failed to commit.*
 
 # --- Sort orders and default-sort-order-id ---
