+++ test/sql/local/catalog_test_config_setup/catalog_agnostic/transactions/test_transactional_updates.test
@@ -77,7 +77,7 @@ select * from my_datalake.default.multi_transaction_update;
 statement error con1
 commit;
 ----
-<REGEX>:.*branch main has changed: expected id.*
+<REGEX>:.*Failed to commit.*
 
 # con1 UPDATE didn't go through
 query II
