+++ test/sql/local/catalog_test_config_setup/catalog_agnostic/transactions/test_transactional_updates_error_omits_stacktrace.test
@@ -69,7 +69,7 @@ select * from my_datalake.default.multi_transaction_update;
 statement error con1
 commit;
 ----
-<REGEX>:.*branch main has changed: expected id.*
+<REGEX>:.*Failed to commit.*
 
 query II
 select * from my_datalake.default.multi_transaction_update;
